Move getThemeFilterToTermTable
