+++ client/state/selectors/test/get-theme-filter-to-term-table.js
@@ -6,7 +6,7 @@ import { expect } from 'chai';
 /**
  * Internal dependencies
  */
-import getThemeFilterToTermTable from 'state/selectors/get-theme-filter-to-term-table';
+import { getThemeFilterToTermTable } from 'state/themes/selectors';
 import { state } from './fixtures/theme-filters';
 
 describe( 'getThemeFilterToTermTable()', () => {
